Рефакторинг OAuth2Service для оптимизации сетевых запросов
