+++ ImageFeed/Network/URLRequestExtension .swift
@@ -1,0 +1,28 @@
+//
+//  URLRequestExtension .swift
+//  ImageFeed
+//
+//  Created by Nataliya MASSOL on 05/10/2023.
+//
+
+import Foundation
+
+// MARK: - HTTP Request
+
+extension URLRequest {
+    static func makeHTTPRequest(
+        path: String,
+        httpMethod: String,
+        baseURLString: String = Constants.defaultApiBaseURLString
+    ) -> URLRequest? {
+        guard
+            let url = URL(string: baseURLString),
+            let baseURL = URL(string: path, relativeTo: url)
+        else { return nil }
+        
+        var request = URLRequest(url: baseURL)
+        request.httpMethod = httpMethod
+        return request
+    }
+}
+
