+++ ImageFeed/Network/URLSessionExtension.swift
@@ -1,0 +1,46 @@
+//
+//  URLSessionExtention.swift
+//  ImageFeed
+//
+//  Created by Nataliya MASSOL on 06/10/2023.
+//
+
+import Foundation
+
+enum NetworkError: Error {
+    case httpStatusCode(Int)
+    case urlRequestError(Error)
+    case urlSessionError
+    case invalidRequest
+}
+
+extension URLSession {
+    func data(
+        for request: URLRequest,
+        completion: @escaping (Result <Data, Error>) -> Void
+    ) -> URLSessionTask {
+        let fulfillCompletion: (Result <Data, Error>) -> Void = { result in
+            DispatchQueue.main.async {
+                completion(result)
+            }
+        }
+        let task = dataTask(with: request) { data, response, error in
+            if let data = data,
+               let response = response,
+               let statusCode = (response as? HTTPURLResponse)?.statusCode
+            {
+                if 200 ..< 300 ~= statusCode {
+                    fulfillCompletion(.success(data))
+                } else {
+                    fulfillCompletion(.failure(NetworkError.httpStatusCode(statusCode)))
+                }
+            } else if let error = error {
+                fulfillCompletion(.failure(NetworkError.urlRequestError(error)))
+            } else {
+                fulfillCompletion(.failure(NetworkError.urlSessionError))
+            }
+        }
+        task.resume()
+        return task
+    }
+}
